feat: extract worlds page component
